A patch from Olivier Heintz - OFBIZ-4506 - BigDecimal issue, when changing quantity to produce for a ProductIonRun
git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1189404 13f79535-47bb-0310-9956-ffa450edef68

+++ applications/manufacturing/src/org/ofbiz/manufacturing/jobshopmgt/ProductionRun.java
@@ -131,7 +131,7 @@ public class ProductionRun {
             try {
                 productionRun.store();
                 if (quantityIsUpdated) {
-                    productionRunProduct.set("estimatedQuantity",this.quantity);
+                    productionRunProduct.set("estimatedQuantity",this.quantity.doubleValue());
                     productionRunProduct.store();
                     quantityIsUpdated = false;
                 }
@@ -204,7 +204,7 @@ public class ProductionRun {
         for (Iterator<GenericValue> iter = productionRunComponents.iterator(); iter.hasNext();) {
             GenericValue component = iter.next();
             componentQuantity = component.getBigDecimal("estimatedQuantity");
-            component.set("estimatedQuantity", componentQuantity.divide(previousQuantity, 10, BigDecimal.ROUND_HALF_UP).multiply(newQuantity));
+            component.set("estimatedQuantity", componentQuantity.divide(previousQuantity, 10, BigDecimal.ROUND_HALF_UP).multiply(newQuantity).doubleValue());
         }
     }
     /**
